Better evauluation of DOS_3d

+++ cpp/HubbardMeanField/sources/Hubbard/DensityOfStates/Square.cpp
@@ -13,9 +13,9 @@ namespace Hubbard::DensityOfStates {
 
 		for (int g = 0; g < Constants::BASIS_SIZE; ++g)
 		{
-			const double gamma = (0.5 + g) * step;
+			const long double gamma = (0.5 + g) * step;
 			values[g] = (M_1_PI * M_1_PI)
-				* boost::math::ellint_1(sqrt(1. - (gamma * gamma / 4.)));
+				* boost::math::ellint_1(sqrt_1_minus_x_squared(0.5 * gamma));
 		}
 		symmetrizeVector<false>(values);
 		computed = true;
